Change message processor Fix memory leaks Increase warning level
git-svn-id: file:///.data/svn_convert/svn@2627 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/plugins/icq/aiminfo.cpp
@@ -121,7 +121,7 @@ void AIMInfo::fill()
     setText(edtCity, data->City);
     setText(edtState, data->State);
     setText(edtZip, data->Zip);
-    initCombo(cmbCountry, data->Country, getCountries(), true, getCountryCodes());
+    initCombo(cmbCountry, (unsigned short)(data->Country), getCountries(), true, getCountryCodes());
 
     if (m_data == NULL){
         if (edtFirst->text().isEmpty())
